Update: refacter nested navgraphs
Previously, the composables within the nested navgraphs were attempting
to navigate directly to each other on starting the app. This cause an
error where the destination was not being found. This has been changed
to a loading screen when the app starts, as the login status of the user
if being checked. if they are authenticated, the navhost will navigate
to the farmerScreens subgraph, otherwise to the authScreens subgraph