Add: NavHostController for NavHost in MainScreen
Previosuly, the navController instantiated in the MainActivity and
passed down was resulting in the error, "ViewModel should be set before
setGraph call". This error shows that you're tring to set the navigation
graph before associating a ViewModelStore with your NavController. The
error has been partly resolved by instantiating a NavHostController for
the NavHost in MainScreen. It is recommended that each NavHost should
have its own NavHostController